[RocksDB] [Performance Branch] Trivia build fix
Summary: make release complains signed unsigned comparison.

Test Plan: make release

Reviewers: kailiu

CC: leveldb

Differential Revision: https://reviews.facebook.net/D14661

+++ util/cache_test.cc
@@ -128,7 +128,7 @@ TEST(CacheTest, UsageTest) {
   }
 
   // make sure the cache will be overloaded
-  for (int i = 1; i < kCapacity; ++i) {
+  for (uint64_t i = 1; i < kCapacity; ++i) {
     auto key = std::to_string(i);
     cache->Release(
         cache->Insert(key, (void*)value, key.size() + 5, dumbDeleter)
